refactor + add guided_choice
