+++ requirements.txt
@@ -12,4 +12,4 @@ pydantic >= 2.0  # Required for OpenAI server.
 aioprometheus[starlette]
 pynvml == 11.5.0
 triton >= 2.1.0
-outlines == 0.0.27
+outlines >= 0.0.27
